website code
This code is written in python. It works with the help of streamlit application which enables a user free smooth interface without any use of front end, middle ware and backend code. You can run it uing the streamlit application using the following command in the terminal "streamlit run /project/location/path.py" and view the results