Add files via upload
This is the model working code. This code is completely written in python. This code contains data loading, preprocessing splitting, training, testing, evaluating and visualizing the result. 

NOTE: I suggest you to run this code in google colab so that would be easy to identify mistakes and execute smoothly. To run this code in colab notebook, you need crop leaf images dataset. It will be attached alongwise