Crop working model.ipynb
This is the code containing the working logic of the deep learning like models VGG16, CNN and Resnet.This code needs to be runned only in google colab. After running the entire code, the model will be downloaded from the code in .keras file format into the system. This file has to be integrated in the below crop website code for predicting the crop diseases and suggesting crops