Add files via upload
NOTE: This code has to be executed only in google colab for making model work. 

This is the actual working file of the model that is executed by me. This code is written in python language for easy coding and simple understanding. When you execute the code in model source code file, the results were observed as seen in this model code file.