Code along with working proof
This is the python code of the model along with the working proof done by me. This code contains training the model with rawdata, loading the model, evaluate and visualize the predicions.

NOTE: This code should be runned only in google colab or jupyter notebook for executing it's results. Try in google colab for better experience and faster executions