TFLite Model
This is not the actual model, but the compressed model file as github do not accept files larger than 25mb. This file is uploaded for your reference only. You can work with the original .keras file that can be downloaded by running the entire model source code file in google colab or Jupyter notebook